Bug 85721
Multiple tooltip shown when hover over message subject contains URL


Undo bug 76098 - do not process Zimlets in the subject line. Another weird case Josh showed me is if there's an email in the subject and it's turned into an ugly (in this case) bubble! :)
Well this means it was no longer consistent with message view, but there also we do not want to process Zimlets objects. It can just get too weird and there's not a lot to gain. So I removed that too.

http://bugzilla.zimbra.com/show_bug.cgi?id=85721