bug:86116
calendar-multiget triggers 500 error and StringIndexOutOfBoundsException

A beta of SimpleCal does a calendar multi-get including the calendar URL as one of the calendar items but we expect the URLs to be for calendar items ending in ".ics".
Now detect that URL doesn't fit the pattern before processing it and ignore it.
With this fix, SimpleCal can at least load a simple calendar.
https://reviewboard.eng.zimbra.com/r/139/
IM ---> main

http://bugzilla.zimbra.com/show_bug.cgi?id=86116