Bug 79342
After saving draft, change button from Cancel to Close

Change the button to "close" after manual "save draft". And also setting it to "close" when editing an existing draft. Both are explicit actions of the user to start editing a "draft", so it makes sense.

Once a compose session changes the button to "close" I keep it simple and keep it "close". Otherwise it would be too confusing I think.

http://bugzilla.zimbra.com/show_bug.cgi?id=79342