Bug 85965
After calendar tag or saved search clicked, can't go back to full calendar view

So I noticed that the "refresh" button does not deselect the current selection from the overview (despite going back to the full all calendars events). This is not good since the tag or search could still be highlighted, confusing the user.

Took me a while to find a way (one way... there might be better) to deselect so there's no confusion.

http://bugzilla.zimbra.com/show_bug.cgi?id=85965