Bug 74541
Folder selection lost after renaming it


Note that this didn't happen if it's an "only child" folder. It happened because rename can cause the folder to move up or down in the children of the parent, so we remove it and get the sort order and re-insert it, in the process losing selection. This fix remembers the selection and re-selects it.

http://bugzilla.zimbra.com/show_bug.cgi?id=74541