bug: 85342
Add validateBlob() to ExternalMailboxBlob and call it before returning from getMailboxBlob(). This ensures that a MailboxBlob instance is always valid; which is the assumption used by most of the code since that is how FileBlobStore behaves.

Also add warning log to POP3Handler so it is more clear when this problem occurs.

http://bugzilla.zimbra.com/show_bug.cgi?id=85342