Bug:84871
Offline setting will be enabled depending on zimbraPrefWebClientOfflineAccessEnabled preference and zimbraPrefWebClientOfflineBrowserKey list. Unique browser key is stored in browser’s localstorage and it is checked with the list of zimbraPrefWebClientOfflineBrowserKey for enabling offline mode.

Once the user disables offline mode manifest file has been regenerated for removing the already stored files in the application cache. IndexedDB data is also removed.

http://bugzilla.zimbra.com/show_bug.cgi?id=84871