Bug 86402
Filter state not maintained between Query Language and Filter Checkbox on left


This makes a connection from the query to the basic filters in the filter panel. So far we only had the reverse connection, from the filters to the query...

This solves not just the case of removing a bubble, but also adding it (it was also not affecting the filter panel).

The fix is not just for "has:attachment", but also is:unread and is:flagged.

Coincidentally this also simply solves bug 80827.

http://bugzilla.zimbra.com/show_bug.cgi?id=86402