Bug: 60017 - Move email zimlet code to core code base
Part Two

Move basic email zimlet functionality into an internal
object handler. It gets a reference to a controller through
the view that contains the object, and uses that to borrow
the bubble action menu as its action menu.

The email zimlet (com_zimbra_email) now just supports a
fancier tooltip for email addresses. (It’s called EmailTooltipZimlet,
so now that’s what it is.) It no longer tries to match content
for email addresses. Its only connection to objects is listening
for zimlet notifications. Removed a bunch of code, but left the
presence stuff for reference.

- remember to handle mailto: links
- when creating filter, use address (rather than msg)
- restore a couple of zimlet notifications that other zimlets
listen for (onMsgView, onFindMsgObjects)

http://bugzilla.zimbra.com/show_bug.cgi?id=60017