bug: 86430
NPE on bootStrap for SimpleCal.
Note.  Apple Mac OSX Calendar auto discovery still doesn't work because it uses the name before the @ in the email address as the username.  The standards suggest if that fails, it should try the full address but that doesn't happen.
The main fix is in Principal class.  It appears that at one point we wanted to use the name without "@" if this was the Zimbra default domain name, but the code was buggy and other places where similar code was put in have been replaced with code that always uses the full address (those call sites are more likely to be used in the real world too, apart from during bootstrap)
User was assuming the current URL was the Principal URL - not true during bootstrap.
DavServlet now makes it clear when sendError fails because it had previously been called.  Was seeing some very confusing logging.
ZimbraServlet now returns null if it calls req.sendError - otherwise we do loads of work and then throw that work away (whilst doing confusing logging...)
At some point, would be good to use AuthUtil for Basic Auth instead but need to check there aren't any important differences.
Suppressed stack trace in W3cDomUtil.
https://reviewboard.eng.zimbra.com/r/280/
IM ---> main

http://bugzilla.zimbra.com/show_bug.cgi?id=86430