Bug: 86606 - JS error on interacting with address
The "Attach Contacts" zimlet tries to add to the action menu for
the "Email Contact Details" zimlet and ran into an NPE. Remove that
code for now. We can add it back when we move the attachcontacts
functionality into core code.

See https://bugzilla.zimbra.com/show_bug.cgi?id=3905#c26 for when
the two zimlets got hooked up.

http://bugzilla.zimbra.com/show_bug.cgi?id=86606