Bug 86556
"send & archive" button should be disabled when uploading attachment


added zimlet notification "enableComposeToolbarButtons" that is called whenever ZmComposeView.prototype.enableToolbarButtons is called, and looks for the send+archive button and enables (or disables) it.

Also did some refactor of the code of enableToolbarButtons, and cleanup of OP/button IDs in the zimlet.

http://bugzilla.zimbra.com/show_bug.cgi?id=86556