Bug: 82441 - Random email addresses used when "Group By: From" is used
I was unable to replicate the bug, either with my dev data or the included messages.  The code is assuming that the first participant is always the 'From' person; my best guess is that this is sometimes an incorrect assumption, so I've altered the code to explicitly look for the participant type, and use that participant.

http://bugzilla.zimbra.com/show_bug.cgi?id=82441