Bug 85261
Intermittently list view goes blank on hitting Esc key after cancelling reply/forward to message


Investigating this seemingly minor bug opened a pandora box of problems. Took me a while to figure out what is the real problem, as I tried all kinds of false leads or ways to hack the problem even more.

A few bugs are involved in the history:

Bug 81245 is the root cause - In this bug in ZmConvController.prototype._initializeView, the controller passed to the view was change from "this" (ZmConvController instance, that made sense) to an instance of ZmConvListController - this is ugly and confusing and cause a bug that was hard to debug.

Bug 83244 - this one was where the regression actually happened. Once the controller was (mistakingly) ZmConvListController, the ESC (ZmKeyMap.CANCEL key map) key handling in ZmConvListController.prototype.handleKeyAction shouldn't have assumed standalone in case.

I reverted bug 81245, and then thanks to a later bug 80224 - in which getFoldersToOmit was moved up from ZmConvListController to ZmMailListController - the problem was is no longer existent.

I also revert bug 83244 fix as it was wrongly implemented. That bug no longer occur with the reverting of bug 81245.

Note that I discovered a few more bugs that this fixes, including:
1. if there was a inline reply open, the ESC would still have closed the standalone tab without warning the user.
2. might also fix a problem with the ZmKeyMap.SEND shortcut, whatever that is.

http://bugzilla.zimbra.com/show_bug.cgi?id=85261