Bug 82353
linkedin preferences: 'undefined' shown


Don't do anything on double click. Set that to do the same as single click, which also does nothing.
The reason it was showing the weird "preferences" is that double click opens some kind of weird default generated preferences from ZmZimletBase.prototype.doubleClicked (ZmZimletBase is the base class of all the zimlets). So there it finds the "userProperty" linkedinZimlet_api_key, but that's not a real user property, we just use this hook to set up the Zimlet with the LinkedIn app key. Maybe there's a better way to do that. But for now this fix is more than enough. It's fast and this bug is very minor anyway.

http://bugzilla.zimbra.com/show_bug.cgi?id=82353