Bug:86854
Web client offline mode supports only message view. To achieve this view is switched to message if the current view is conversation resulting in this bug. For web client offline mode search type is changed to message, which will always result in message view. We are following the same approach for drafts folder. 

http://bugzilla.zimbra.com/show_bug.cgi?id=86854