bug: 86850
ImapSessionManager$SessionSerializerTask deadlock.
Use ConcurrentLinkedHashMap instead of LinkedHashMap.  This allows us to iterate over the sessions in an ordered fashion without taking a lock on the map which exposed us to the danger of deadlock and should also reduce contention for the map.
Used to iterate over all sessions from least recently accessed to most recently accessed.  Reversed this as we are more likely to evict items later in the processing.
Used to use null as the value in entries but that is not allowed in the new map, so re-use the key as the value.
https://reviewboard.eng.zimbra.com/r/389/
IM ---> main

http://bugzilla.zimbra.com/show_bug.cgi?id=86850