Bug 87002
Exception when switching reading pane view after clicking on an invite and a non-invite


The reason is inconsistency - we would have an inviteMsgView but the message would not have an _invite.

I found that this is due to ZmMailMsgView.prototype.reset, the this._inviteMsgView is reset() but not nulled. So I fixed that.

http://bugzilla.zimbra.com/show_bug.cgi?id=87002