Bug 86998
Refactor/Cleanup the double pane sash move code


Refactor/cleanup of the ZmDoublePaneView.prototype._sashCallback method.

A couple of the main things:
(a) There's really no reason to distinguish so much between negative and positive delta
(b) I also found a way to calculate the actual delta (based on minimum width/hight of views) in a common way for reading pane on right and on bottom.

This code change also fixes 2 subtle and minor bugs:
1. moving the sash UP fast, would cause it to not move as up as it could.

2. Also in "reading pane at bottom" - if the first list view the user looks at is empty, we allow the user to move the sash up all the way to the header instead of 2 rows below it, even after switching to a view with a non empty list.

http://bugzilla.zimbra.com/show_bug.cgi?id=86998