Bug 87002
Exception when switching reading pane view after clicking on an invite and a non-invite


Re-opened - somehow my seemingly innocent fix caused a regression that caused the message view (regular, after clicking on invite one) to stay the width that it was when the "day view" was there in the invite message.

The problem was that due to completely nulling the _inviteMsgView, the resize(true) was not called.

So I first added a call in ZmMailMsgView.prototype.reset to
this._inviteMsgView.resize(true);

That seemed to work, but then I thought maybe we'll got some other regressions, and also noticed that at least initially (according to the comment) ZmInviteMsgView was created as a Singleton, so I though, sure, let's not null it but make sure this works otherwise.

So what I did is fix it in ZmMailMsgView.prototype.setVisible  - if dayView is null, this is not really an invite so return fast.

http://bugzilla.zimbra.com/show_bug.cgi?id=87002