bug 32283 - accessibility: use Dwt.byClassName() and Dwt.byTag()
Replace skin.byClass() with Dwt.byClassName(), and some references to
getElementsByTagName() with Dwt.byTag(). As both of the Dwt functions
return NodeLists, fix them to return arrays, and make
AjxUtil.toArray() capable of performing the conversion. The latter
involves recognizing array-like elements, so we add a function that
does so -- AjxUtil.isArrayLike().

http://bugzilla.zimbra.com/show_bug.cgi?id=32283