bug 32283 - fix HTML editor focus & tabbing
This fixes having an editor in the tab order, and allows the editor to
receive focus regardless of whether it's in HTML or text mode. In
addition, we add the toolbar to the tab order after the editor field
itself, and apply two patches to TinyMCE: one to allow all toolbar
buttons to receive focus, and one to allow the tab order to continue
from the toolbar.

Please note that although we apply these changes to the stock classes,
the toolbar won't appear in the tab order there, as users cannot tab
out of the editor.

http://bugzilla.zimbra.com/show_bug.cgi?id=32283