Bug 86846
vertical scroll bar missing in conv view


UPDATE to previous fix - remove the 100% for IE10 and IE9 as well.

Probably regression from bug 83487.

Padding was added (in skin.css) to class .Conv2Messages. (There's just one element of this class, it's the messages part in the preview pane, not including the header). Not sure how it was before, looks like there was no padding to that particular element but still it looked fine (as in Dogfood).

Problem with padding is that in combination with it,  width=100% of .Conv2Messages (defined for a long time in zm.css) doesn't work quite right, I think it doesn't take the padding into consideration and sets the width without the padding to the parent's width (though it does not look like it does that exactly either). See this:
http://stackoverflow.com/questions/5219175/width-100-padding
(explanations and also possible solutions if this 100% is really needed).

The width 100% was added in bug 67841, and that bug is about IE...now I do this only if it's MSIE_LOWER_THAN_9

http://bugzilla.zimbra.com/show_bug.cgi?id=86846