Bug: 86178 - File not repositioned after rename
Also, found and fixed a problem when adding a file (via upload dialog or drag and drop) - it can be placed incorrectly in the list if any item with older versions has been expanded (whether left open or collapsed).  The view's list contains only file entries, while the display list contains file entries and versions.  New entries were inserted correctly into the list, but in the wrong spot in the display list.

http://bugzilla.zimbra.com/show_bug.cgi?id=86178