Bug 87174
"delete" button is mis aligned


The problem was indeed caused by the archive zimlet. There the code decided whether to show or hide the "Delete" button. (based on the user's zimlet pref). However, when displaying the "Delete" button - it chose the display style of "inline-block". My guess is that it was the case that this was the style, when the zimlet was written, but changed.

Anyway my fix is to store the display style when hiding, so when displaying it again, we can use it. (note that we really re-use it only if the user switches from "hide" to "don't hide" the delete button, but this fix takes care of this case, and also of the case the button is simply displayed all the time to begin with, as it then sets it to the same value). Definitely, setting it to "inline-block" (of even changing it to "block" or "" as it would be good now) is not the right approach as it would break if we change it again.

http://bugzilla.zimbra.com/show_bug.cgi?id=87174