Bug 87499
Expanded conv message list does not get messages removed on delete


I'm not sure anymore if historically we showed the expanded conv list view with strikethrough when deleted (moved to trash more precisely). Anyway I fixed the "moved to trash" part relatively easily in ZmConvListView.prototype._changeListener, to make it behave like spam and "delete" (at first I thought the correct place would be ZmConv.prototype._msgListChangeListener, as it seems very specifically to changes in that expanded conv, but then saw that for spam, and hard delete, it's dealing with those in ZmConvListView.prototype._changeListener so that's where I made the change)

I then though noticed that undoing the delete does not return the messages to the expanded conv (this is true for undoing spam too). I guess it's not that important but I spent some time trying to see if I can do it and I do believe I found a reasonable solution to this.

http://bugzilla.zimbra.com/show_bug.cgi?id=87499