Bug: 87635 - Conversation shows empty msg headers
Somehow the dogfood instance of the touch client lacks
ZCS.constants.APP_NAMES. Use the app array instead.

Add a null check to ZtItemController::getItemButton()

Instead of adding a function to XTemplate to enable nesting,
just create a function to make it easy to include the output
of one template within another.

http://bugzilla.zimbra.com/show_bug.cgi?id=87635