Bug 85962
Mail Filter rule for social twitter does not work


Discussed with Josh. We never really implemented those 2 "@ direct messages and mentions" options correctly for any of the social networks. The worst part is that negative meant not just say Twitter that's NOT DM or mention, but also anything else (some mail from foobar.com). There was never (I think never) any 2 part thing like "Twitter DM or mentions" vs. "twitter in general that's not DM or mentions". It was just Twitter DM and mentions vs. everything else from twitter and from the rest of the world. And now we can't really detect DM or mentions (this is this bug).

So what I did is removed the value dropdown. And modified the text to be more specific to what we try to filter.

Instead of Facebook, the dropdown will say "Facebook notifications"
Instead of LinkedIn, the dropdown will say "LinkedIn messages and connections"
Instead of Twitter, the dropdown will say "Twitter notifications"

I also enhanced the server side filter of LinkedIn just a bit, as I noticed I get a lot of LinkedIn messages from the address hit-reply@linkedin.com, so added this to the other 2 addresses already checked for.

http://bugzilla.zimbra.com/show_bug.cgi?id=85962