Bug 81015
After setting "Mark messages in reading pane as read after" to a value, the mail will be marked as read, after clicking on the message, then clicking on another message etcetera after the define amount


ZmConvListController.prototype._displayItem overrides ZmDoublePaneController.prototype._displayItem but does not call killMarkReadTimer unlike the method it overrides. It is overridden to do some stuff a little differently due to some marking read issues. But this fix is simply to add that killMarkReadTimer call (that I extracted to a method in bug 68886, luckily).

http://bugzilla.zimbra.com/show_bug.cgi?id=81015