bug:87718
MailboxIndex:
Mitigate one of the effects of corrupt metadata column in the database - a user cannot select a folder containing such an item.
Discard any items from the results that cannot be constructed from their underlying data.

Metadata:
In some instances, it appears that an encoded string is getting corrupted by being treated at some point as if the bytes were ISO-8859-1 instead of UTF-8.  Try again with this reversed.  If this works, the internal lengths will have gotten corrected, so there is a fair chance this has correctly identified what happened.
https://reviewboard.eng.zimbra.com/r/538/

http://bugzilla.zimbra.com/show_bug.cgi?id=87718