bug:85111
Like with "FN", make an effort to make sure "N" has a meaningful value.
Fix mainly done for groups which seem fairly hiden in Mac OSX Maverick Contacts app without "N".  This actually applies to ordinary contacts too (to a slightly lesser extent - some fields like email will be used in the absence of "N")

Worth testing creating groups and contacts in ZWC with ";" (and perhaps other) characters in parts of the name - these should be escaped in the first component of "N" if it is based on the value of "FN".
https://reviewboard.eng.zimbra.com/r/625/

http://bugzilla.zimbra.com/show_bug.cgi?id=85111