Bug 87929
Marking multiple items as Not Spam doesn't move them out of the Junk folder


Happens with single message too (or rather, conversation, this issue won't occur in message view).

And wow, another crazy === causing a bug! Pretty scary as this is a problem in getTcon and could affect actions other than spam. Glad this bug helped us detect this issue.  I fixed that.

Regression from bug 85084.

the tcon was wrong, containing "j" so the messages were not moved from spam...

I also noticed that for spam action we call the getTcon even for msg view, and there's no reason (we even pass it to the server, that probably just ignores it). I fixed it.

http://bugzilla.zimbra.com/show_bug.cgi?id=87929