Bug 87714
Bottom of msg cut off


First problem: the div did not have a width set (so it was auto or even worse, 1772), so this caused a problem in messages that have long text that wraps (and other things too).

Solution to first problem: This is easy to fix like you suggested by setting the div width to the body width (of the capsule msg). There were some subtle things like taking the padding into consideration.

Second problem: We cache the size by the msg ID. We do not clear it on resize of the view pane or switching views.

Solution to second problem:  Forget about the caching - not sure how helpful it is, what percentage of messages viewed by a user are new and what are ones they already viewed in this session? Probably not big. What are we saving? Probably not much, like  mentioned in comments - since it's only for short messages, it won't be too expensive.

Third problem: When we "bail" in case the body height is bigger than 150, we need to reset the iframe height to auto (or so I thought), in case we already resized it (this is important in case again, the user moves the sash, and makes the reading pane narrower, till the body is higher than 150 - but we already sized the iframe to less than 150.  However, this does not really work well. I now discovered that the iframe is also resized (to be bigger than 150, the default is not to fit the content...) but the following code (From the traditional MSG view) ZmMailMsgView._resetIframeHeight. So my solution is when we bail to also call this on timer (it's called on timer from other places, so just use that method, timer is not really required). No idea why _resetIframeHeight is not already called on sash move...

http://bugzilla.zimbra.com/show_bug.cgi?id=87714