Bug: 85382, 87386, 87943 - Improve error handling
I wasn’t able to get Sencha Touch to call a failure callback
directly on store.load() - the Operation has it, but it’s not
passed down to the Request where it’s looked for. The workaround
is to look for it in the Operation when we’re handling a response
in our ZtSoapProxy.

- Remove “Loading…” popup when an error is detected
- If we can’t connect to server, show an error
- Ignore error if request was a poll

http://bugzilla.zimbra.com/show_bug.cgi?id=85382
http://bugzilla.zimbra.com/show_bug.cgi?id=87386
http://bugzilla.zimbra.com/show_bug.cgi?id=87943